fix(web): catch stale-chunk errors via vite:preloadError + harden guard
Address review of the error-boundary recovery:
- Extract the detect + guarded-reload logic into app/error/chunk-reload.ts so
  the boundary and a new listener share one implementation.
- Add a window "vite:preloadError" listener in entry.client that recovers
  before the error reaches the boundary (Vite's first-class signal for failed
  dynamic imports — more robust than message matching); preventDefault on reload.
- Also match the CSS-preload wording ("Unable to preload CSS for ...").
- Wrap sessionStorage access in try/catch so private-mode/disabled-storage
  browsers still reload instead of throwing past the guard.
