+++ apps/web/app/error/index.tsx
@@ -7,28 +7,11 @@
 import { useEffect, useState } from "react";
 // hooks
 import { useAppRouter } from "@/hooks/use-app-router";
+// utils
+import { isChunkLoadError, reloadOnceForStaleChunk } from "./chunk-reload";
 // layouts
 import { DevErrorComponent } from "./dev";
 import { ProdErrorComponent } from "./prod";
-
-// A stale-deploy failure: the user's cached index.html references a hashed
-// chunk the new build no longer serves, so a lazy() import 404s. Vite/Rollup
-// and each browser word this differently.
-function isChunkLoadError(error: unknown): boolean {
-  const err = error as { name?: string; message?: string } | null;
-  const message = err?.message ?? "";
-  return (
-    err?.name === "ChunkLoadError" ||
-    /failed to fetch dynamically imported module/i.test(message) ||
-    /error loading dynamically imported module/i.test(message) ||
-    /importing a module script failed/i.test(message)
-  );
-}
-
-// Only reload once per short window so a genuinely-missing asset (build/CDN
-// gap, not just stale) falls through to the error screen instead of looping.
-const RELOAD_GUARD_KEY = "pi-dash:chunk-reload-at";
-const RELOAD_GUARD_MS = 10_000;
 
 const handleReload = () => window.location.reload();
 
@@ -37,20 +20,16 @@ export function CustomErrorComponent({ error }: { error: unknown }) {
   const router = useAppRouter();
   // Recover from stale-deploy chunk errors by reloading once. Start in the
   // recovering state so we don't flash the maintenance screen before reloading.
+  // (The vite:preloadError listener in entry.client usually catches these first;
+  // this boundary is the fallback for chunk errors that slip past it.)
   const [recovering, setRecovering] = useState(() => isChunkLoadError(error));
 
   const handleGoHome = () => router.push("/");
 
   useEffect(() => {
     if (!isChunkLoadError(error)) return;
-    const last = Number(sessionStorage.getItem(RELOAD_GUARD_KEY) ?? 0);
-    if (Date.now() - last > RELOAD_GUARD_MS) {
-      sessionStorage.setItem(RELOAD_GUARD_KEY, String(Date.now()));
-      window.location.reload();
-    } else {
-      // Already tried recently — the asset is likely gone, show the real error.
-      setRecovering(false);
-    }
+    // If we reloaded too recently, the asset is likely gone — show the real error.
+    if (!reloadOnceForStaleChunk()) setRecovering(false);
   }, [error]);
 
   if (import.meta.env.DEV) {
